curl: Autospec creation for version 7.55.1
Add a patch to avoid running a configure routine that strips the -g
compiler flag from CFLAGS unless --enable-debug is specified.

As noted in the patch description, --enable-debug appears to be more
oriented for curl developer use. We simply want -g to be passed to gcc
so that complete debuginfo is generated for the build.

Signed-off-by: Patrick McCarty <patrick.mccarty@intel.com>